+++ src/lib/mrvCore/mrvHotkey.h
@@ -180,6 +180,7 @@ namespace mrv
     
     extern Hotkey kVideoLevelsFile;
     extern Hotkey kVideoLevelsLegalRange;
+    extern Hotkey kVideoLevelsLegalRangeHDR;
     extern Hotkey kVideoLevelsFullRange;
 
     extern Hotkey kAlphaBlendNone;