+++ src/lib/mrvUI/mrvMenus.cpp
@@ -792,6 +792,17 @@ namespace mrv
                 timeline::InputVideoLevels::LegalRange)
                 item->set();
 
+#ifdef VULKAN_BACKEND
+            idx = menu->add(
+                _("Render/Video Levels/Legal Range HDR"),
+                kVideoLevelsLegalRangeHDR.hotkey(),
+                (Fl_Callback*)video_levels_legal_range_hdr_cb, ui, mode);
+            item = (Fl_Menu_Item*)&(menu->menu()[idx]);
+            if (imageOptions.videoLevels ==
+                timeline::InputVideoLevels::LegalRangeHDR)
+                item->set();
+#endif
+            
             idx = menu->add(
                 _("Render/Video Levels/Full Range"),
                 kVideoLevelsFullRange.hotkey(),