+++ tlRender/lib/tlCore/Image.h
@@ -140,6 +140,7 @@ namespace tl
         enum class VideoLevels {
             FullRange,
             LegalRange,
+            LegalRangeHDR,
 
             Count,
             First = FullRange