+++ tlRender/lib/tlTimeline/ImageOptions.h
@@ -21,6 +21,7 @@ namespace tl
             FromFile,
             FullRange,
             LegalRange,
+            LegalRangeHDR,
 
             Count,
             First = FromFile